+++ src/ports/study/index.ts
@@ -1,0 +1,12 @@
+export {
+  IdempotencyConflictError,
+  createAnswerSubmissionFingerprint,
+  type ClockPort,
+  type CommitAnswerInput,
+  type CommitAnswerResult,
+  type IdGeneratorPort,
+  type LearningEventRepositoryPort,
+  type LearningTransactionPort,
+  type StudyPersistencePort,
+  type StudySessionRepositoryPort,
+} from './StudyPersistence';